Add color package to dependencies
This adds the color (https://www.npmjs.com/package/color) package as a
dependency. The color package allows us to manipulate colors and
generate corresponding css strings in a sane manner. This makes mixing
and adjusting colors easy, and avoids stupid and complex string logic
when it comes time to generate style strings to insert into the markup.