Create LEDContainer component
This adds and LEDContainer component. This component's purpose is to
serve to render individual LEDs based on a state object that is passed
to the container. This container is responsible for rendering the
style of the LEDs as well as their position, though perhaps the styling
logic should be moved down into the LEDs themselves.