Add LEDContainer to app
This adds the LEDContainer element to the app.js, as well as some css
styling to make it behave how we want.

For testing I'm including a few LEDs of different sizes and at random
locations, but in the future these LEDs will be fed from some sort of
input data stream.