Move style logic down to LED
In the interests of encapsulation, I've moved the logic for actually
computing the styles (position, size, color, boxShadow) down into the
LED, because it feels like ultimately the LEDContainer should really
just be that: a container.

In doing this, we update the array that is feeding the LEDs in App.js
to have a position property.