Updated box shadow rendering
This updates the calculation of the box shadow string to make one of
the shadows size dependant on the size of the LED. This gives a more
pleasing appearance, but the appearance of the LEDs still isn't quite
there.

We also alter the color of the blue test LED to be lighter, so that
it's easier to compare to the reference implementation of CSS LEDs that
I'm going off of.