Move leds to state and animate
This is a first stab at seeing whether animating the LEDs is viable.
Based on this I would say this is a success.

Firstly we had to move the LED data to the main App's state, because
the props are not mutable where as the state is. Secondly, in order to
make use of some of the more advanced and powerful lifecycle features
of the React Component spec, we needed to change from creating the
component by extending Component to calling React.createClass.
Lastly, we included some methods to incrementally change the color of
the LEDs, and the code needed to update the internal state to make this
change render.