Add an LEDLayout module for arranging leds
This adds the LEDLayout module which is intended to offer helper
functions that assist with laying out LEDs in 2D space according to
different layout methods. The intent behind this module is to
encapsulate layout logic in a single place, with a single uniform
interface.

The interface that should be provided by functions exported from the
LEDLayout module is that the functions should be passed the total
number of LEDs and the ids of each. At a later stage it may make more
sense to just pass the total number of LEDs and return an array of
locations. The functions should return the corresponding position of
the LED in the range of -1 to 1 in both axis. Again, this may be
changed to some other coordinate system at a later stage.

Currently the module only provides on layout type, concentric layout.
This arranges the LEDs in concentric circles with a predetermined
number of LEDs in each circle. LEDs are laid out so as to be evenly
spaced within their circle.