Make use of LEDLayout module in arranging LEDs
This makes the LEDContainer component use the new layout module to lay
out LEDs. In order to test this properly, we generate LEDs at random
instead of via a fixed array, and we can alter the number of LEDs in
order to test the layout function.

We also modify the CSS slightly to make the LEDs lay out with absolute
location relative to the relative location of their parent container.
This ensures that they arrange themselves as we want to. We use
percentage based top and left attributes to ensure that the arrangement
scales to the size of the containing div.