Change LEDLayout coordinate system and return type
This changes the way in which methods inside the LEDLayout module are
expected to return their data, and the coordinate system for the data
they return.

The functions in this module should now be called by just passing in
the total number of LEDs that need laying out, and the functions should
return an ordered array with the positions of the LEDs.

These positions now give coordinates in the range 0 to 1 in both axis.
This makes it much easier to convert to percentages when we're feeding
the top and left values to the LEDs.