Add container component to hold graph related bits
This adds a container component that will be used to hold the graph
related components and mediate them.

Currently we have some state that
determines which bits of data the graph detail viewer should show, and
these can be edited using the react browser tools to see what happens
when we change these. Soon these will be wired up to a wider graph view
with some handles that will change these values. For the moment though,
changing things through the dev tools works, but we may have to tweak
the animation speed to make things more fluid, as it's currently far
too slow.