Add component to render led update textbox
This adds a first pass of a component that will be used to update the
number of leds. We render a simple textbox and an update button. Had to
figure out how to allow the user to actually enter text in the textbox,
because by default, it's locked to the value in state.