Add wiring to enable update button functionality
This adds a property on the ledTextbox that accepts a function which
will be called with the new number of leds to display. As part of this,
we rewrite our parsing behaviour to parse the value that has been
entered as an integer, and display an error message (via a state field)
if what was entered couldn't be parsed.

Currently the parsing method is no good (will parse things like 20abc
as 20), so we need to explore other options for integer parsing.