Add select box that can update the LED layout
This adds a select element that can be used to set an autolayout of the
leds in the viewport. We do this by providing currently a grid and a
circle layout, and using layout methods to arrange the leds according
to the pattern.

Currently we need to make this dynamic based on the number of leds, at
the moment we just use a static number because I was in a rush while
streaming.

Also updated our tsconfig to make use of the latest ES features,
particularly Object.entries