Update number of LEDs to obey state
This updates the number of LEDs to obey the size of the current
LEDArray parameter when changing layout. Colors of leds are still
randomized, due to the fact that mapping the positions to the
existing leds is probably not worth it, due to the format of data that
we will be receiving.