WIP: Adding Redux
This is a WIP commit. I've tried to add redux to track state, because I
found that I had two actions that needed to do basically the same
thing, hence, redux. Got stuck when trying to pass null/undefined to
the connect function when connecting up our map state to props and map
dispatch to props, because some of our components don't need mapped
state or dispatch.