Add redux state management using react-redux
This adds state management of the led portion as well as their layout
and number using react-redux. Lots of help from the typescript react
starter page on this one. Still need to fix up adding the random layout
as an option and as the default in the dropdown list.

Note that the issue from the previous WIP commit where something was
wrong with connect was caused (I think) by the fact that since the map
functions use arrow function syntax, the last argument is auto returned
(? maybe). So what the typechecker was complaining about is that for
example updateNumberOfLeds is actually of type (value: number) =>
UpdateNumberOfLeds (our type of the update action). Adding this type as
the return type of the corresponding prop on the component (instead of
void) solved any type checking issues.