Add reselect library to avoid storing derived state
Previously the array of leds was stored in state. However, since this
is a derived value, there is no need to store it. The reselect library
provides a mechanism to generate a memoized version of the function
that generates the derived state from it's dependencies, and we
implement that here to grant us a cleaner design.