Refactor to separate led and data reducers
This refactors the Redux store to have separate reducers for actions
and data relating to the led state and the graph state.

The graph state is newly added (that's what prompted to splitting of
the reducers), and the graph state is now wired (using connect) into
the presentational components surrounding the graphing ie. the graph
start handle. The graph start handle now generates actions which update
the viewport.